refactor(editor): markdown commit helper 분리
