+++ front/e2e/editor-studio-state.spec.ts
@@ -178,6 +178,13 @@ test.describe("editor studio state", () => {
     expect(blockEditorEngineSource).toContain("border-left: 4px solid")
     expect(blockEditorEngineSource).toContain("border-radius: 0;")
     expect(blockEditorEngineSource).toContain('from "./blockSelectionModel"')
+    expect(blockEditorEngineSource).toContain('from "./useBlockEditorMarkdownCommit"')
+    expect(blockEditorEngineSource).not.toContain("const normalizeMarkdown =")
+    expect(blockEditorEngineSource).not.toContain("markdownCommitTimerRef")
+    expect(blockEditorEngineSource).not.toContain("markdownCommitIdleHandleRef")
+    expect(blockEditorEngineSource).not.toContain("markdownCommitMaxWaitTimerRef")
+    expect(blockEditorEngineSource).not.toContain("pendingCommitEditorRef")
+    expect(blockEditorEngineSource).not.toContain("lastCommittedMarkdownRef")
     expect(blockEditorEngineSource).not.toMatch(/const\s+isStableBlockHandleState\s*=/)
     expect(blockEditorEngineSource).not.toMatch(/const\s+isStableBlockSelectionOverlayState\s*=/)
     expect(blockEditorEngineSource).not.toMatch(/const\s+resolveOuterBlockSelectionGesture\s*=/)
